mobile_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-multipart>=0.0.9
